requirements.txt
streamlit==1.27.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pillow>=7.1.0,<10.0.0
scipy==1.10.1
joblib==1.3.2
scikit-surprise==1.1.3 